Added a new scene with a new iteration of the Golisano Track.
I put the new version of the track into a new scene to avoid messing with anything that GSP put into place to handle loading scenes, and for archival purposes.
The new version of the track is not ready for gameplay elements yet, it is too rough and keeps launching the kart into the air.