Volume methods added?
Still needed to be used somewhere and sound team may or may not have a better idea.

Also lowered transition time. Old code was somewhat of a screwup as I tried to anticipate how long it would take for the animation of the start and end total to finish when I can just add add a yield in the middle and trigger the animations separately with different triggers?