issues fixed and discovered
The new buttons are back in. I had most of my work done on a faulty project so I copied and renamed the scene and meta file. The buttons were the only change so I moved the buttons out of the canvas before copying the canvas over.

Anyways I also discovered that the kart model appears in front of the options. Will solve this at a future date, probably.