kart model now moves better with screen size changes
Why did I spend my time doing this? It is not like the canvas elements change size if we squish it on the x-axis.

The character buttons not aligning with the orange stripe might be able to be solved if I get it as a separate image, not sure how high priority that is.

Anyways the kart model now moves better with the canvas if you drag it up, down, left or right