Fixed a selectability issue with mouse and controller at the same time
Mouse click that arent on a selectable ui element does not unselect/hover the last selected anymore.

I changed the StandaloneInputSystem component in the EventSystem object in all the menu scenes to InputSystemUIInuptModule and toggled Deselect On Background Click Off.

Although I think the better feeling interaction would be to jump into the button set rather than to never leave it.