Character selection works!
Added a few new features. One is that you can now see the actual model appear in the Char Selec screen. You can also rotate the model with the kart. When going into a map scene it detects which model it should be having and deletes the others from existing.  To do this I had to edit a few scripts