Minor Changes (Read Description)
Because names are set to lower when checking which model to set active, it cannot find OL model since it is uppercase.

Temporary fix to Appearance Settings, now it checks if the kart is the player kart before grabbing CharacterData information. NPC kart icons are currently manually set in scene.

Applied Jaden's changes to the Pause prefab because the changes were only made to the instances in the scenes itself.