more changes than I thought (read description)
Animator is also stopped during timescale = 0 so I had to make that use unscaled time. Animator is located in StartScene

SpeedCamera does not take effect during timescale = 0 so I manually aligned its view to the CameraFollowFront of the player kart

bool is now in pausehandler instead of countdown and countdown has a reference to the pausepanel.

Because it is using real time if there is lag or something then the countdown looks off.