Checking in
PauseHandler and Countdown now uses Time.timeScale to pause stuff that uses Scaled time which most of the game stuff

I set animator to use unscaled time or the fade out will not fade out until after the countdown is finsihed. The animator is located in the start scene.

I have scene changes because I decided to hook up the countdown script to the pausehandler script through the inspector it should be less work for the computer than having to use the FindAnyObjectByType function everytime pause is called