Solving issue #410
The previous issue was that because it was a break line instead of a continue the loop stopped early before destroying the jamster model since it is futher down the list than other models.

By default the jamster model is active so it is noticeable when it is not destroyed. When other models are not destroyed you can't really tell since they are inactive by default.