Attempting to fix pausing issue
Issue in question: trying to use LoadMap or LoadScene when paused in a map causes a black screen until you unpause and then the scene loads.
Created a DeactivatePause method that successfully deactivates the pause panel when loading into a new scene or map, but it is not registering in the same way that the command panel does when activating/deactivating, so you have to pause and unpause using the keybind again to properly load into a new scene/map. I commented out where I called my method simply so you only have to press esc once to solve the problem instead of twice (to reactivate and then deactivate pause, instead of just having to deactivate it).