New Item Block Type
Item respawn lowered to 2.5s across the board

Random Respawn Type Item Added (RandomRespawnTypeBrick.prefab)
This block type respawns as a random type when it respawns.

Two togglable Booleans added to the item box scrips as serialized fields
RandomRespawnType - Makes the box come back as a random type excluding the upgrade block.
RandomizeOnStart - Makes the box become a random type at the start of the race. This allows you to set the type of a block at the start of a race or not.

Random Respawn Type Item Box has these two on by default the rest of the block types have them off.