Quick Fix
Upon discussion with Keelon it has come to my attention that removing that part is a mistake unless I also change some other code.

There are multiple solutions, I just chose the one that is the least work among the ones that I know.

Examples:
1. Change scene for host & client with a single rpc call
2. Or ignore host in rpc call if you already made the host change scene
3. My solution is that unless specified do not allow scene to load if already in scene
4. There may or may not be other solutions

This will need to be changed if it has been decided that restart should work across multiplayer and one of the other solutions will be better